Restructure Tokenizer and Splitter modules
